Account for console resizing by adding extra mode
Typing disabled mode added, like a preview mode. The user's focused will be forced on the tree, but the editor will display the note that is too big. Lines that are too long will be truncated with an elipsis, and notes that have too many lines will have the last line replaced with an elipsis. The title of the note will get a message in red notifying them that they must make the screen bigger in order to edit the note.